app/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile app/requirements.ini
#
anyio==3.6.1
    # via
    #   starlette
    #   watchfiles
bcrypt==3.2.2
    # via -r app/requirements.ini
cffi==1.15.1
    # via bcrypt
click==8.1.3
    # via uvicorn
ecdsa==0.18.0
    # via python-jose
fastapi==0.79.0
    # via -r app/requirements.ini
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r app/requirements.ini
h11==0.13.0
    # via uvicorn
httptools==0.4.0
    # via uvicorn
idna==3.3
    # via anyio
psycopg2-binary==2.9.3
    # via -r app/requirements.ini
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.9.2
    # via
    #   -r app/requirements.ini
    #   fastapi
python-dotenv==0.20.0
    # via uvicorn
python-jose==3.3.0
    # via -r app/requirements.ini
python-multipart==0.0.5
    # via -r app/requirements.ini
pyyaml==6.0
    # via uvicorn
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-multipart
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.40
    # via
    #   -r app/requirements.ini
    #   sqlalchemy-utils
sqlalchemy-utils==0.38.3
    # via -r app/requirements.ini
starlette==0.19.1
    # via fastapi
typing-extensions==4.3.0
    # via
    #   pydantic
    #   starlette
uvicorn[standard]==0.18.2
    # via -r app/requirements.ini
uvloop==0.16.0
    # via uvicorn
watchfiles==0.16.1
    # via uvicorn
websockets==10.3
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
